LU-307 Send parent FID from client to MDS on NFS open for striping info
Which is useful for delay created cases. and it is mainly for Lustre re-export
through NFS. If client does not give such hint, MDS will try to find parent by
the first fid in the link EA. Another fixes are about the protection of delay
creating OSTs objects concurrently.

Reviewed-on: http://review.whamcloud.com/557
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>

+++ lustre/include/lustre_mds.h
@@ -87,7 +87,7 @@ int mds_log_op_unlink(struct obd_device *, struct lov_mds_md *, int,
 #define MDD_OBD_UUID     "mdd_obd_uuid"
 #define MDD_OBD_TYPE     "mds"
 
-static inline int md_should_create(__u32 flags)
+static inline int md_should_create(__u64 flags)
 {
        return !(flags & MDS_OPEN_DELAY_CREATE ||
                !(flags & FMODE_WRITE));
